+++ backend/canisters/user/impl/src/updates/c2c_notify_events.rs
@@ -5,7 +5,7 @@ use canister_tracing_macros::trace;
 use types::{Achievement, DiamondMembershipPlanDuration, MessageContentInitial, ReferralStatus, Timestamped};
 use user_canister::c2c_notify_events::{Response::*, *};
 use user_canister::mark_read::ChannelMessagesRead;
-use user_canister::{Event, UserCanisterEvent};
+use user_canister::{LocalUserIndexEvent, UserCanisterEvent};
 
 #[update(guard = "caller_is_local_user_index", msgpack = true)]
 #[trace]
@@ -20,41 +20,41 @@ fn c2c_notify_events_impl(args: Args, state: &mut RuntimeState) -> Response {
     Success
 }
 
-fn process_event(event: Event, state: &mut RuntimeState) {
+fn process_event(event: LocalUserIndexEvent, state: &mut RuntimeState) {
     let now = state.env.now();
 
     match event {
-        Event::UsernameChanged(ev) => {
+        LocalUserIndexEvent::UsernameChanged(ev) => {
             state.data.username = Timestamped::new(ev.username, now);
         }
-        Event::DisplayNameChanged(ev) => {
+        LocalUserIndexEvent::DisplayNameChanged(ev) => {
             state.data.display_name = Timestamped::new(ev.display_name, now);
             state.data.award_achievement_and_notify(Achievement::SetDisplayName, now);
         }
-        Event::PhoneNumberConfirmed(ev) => {
+        LocalUserIndexEvent::PhoneNumberConfirmed(ev) => {
             state.data.phone_is_verified = true;
             state.data.storage_limit = ev.new_storage_limit;
             openchat_bot::send_phone_number_confirmed_bot_message(&ev, state);
         }
-        Event::StorageUpgraded(ev) => {
+        LocalUserIndexEvent::StorageUpgraded(ev) => {
             state.data.storage_limit = ev.new_storage_limit;
             openchat_bot::send_storage_ugraded_bot_message(&ev, state);
         }
-        Event::ReferredUserRegistered(ev) => {
+        LocalUserIndexEvent::ReferredUserRegistered(ev) => {
             state.data.referrals.set_status(ev.user_id, ReferralStatus::Registered, now);
             openchat_bot::send_referred_user_joined_message(ev.user_id, ev.username, state);
         }
-        Event::UserSuspended(ev) => {
+        LocalUserIndexEvent::UserSuspended(ev) => {
             openchat_bot::send_user_suspended_message(&ev, state);
         }
-        Event::OpenChatBotMessage(content) => {
+        LocalUserIndexEvent::OpenChatBotMessage(content) => {
             let initial_content: MessageContentInitial = (*content).into();
             openchat_bot::send_message(initial_content.into(), Vec::new(), false, state);
         }
-        Event::OpenChatBotMessageV2(message) => {
+        LocalUserIndexEvent::OpenChatBotMessageV2(message) => {
             openchat_bot::send_message(message.content.into(), message.mentioned, false, state);
         }
-        Event::UserJoinedGroup(ev) => {
+        LocalUserIndexEvent::UserJoinedGroup(ev) => {
             // Check that the user didn't already leave the group before this event arrived
             if !state
                 .data
@@ -71,7 +71,7 @@ fn process_event(event: Event, state: &mut RuntimeState) {
                 state.data.award_achievement_and_notify(Achievement::JoinedGroup, now);
             }
         }
-        Event::UserJoinedCommunityOrChannel(ev) => {
+        LocalUserIndexEvent::UserJoinedCommunityOrChannel(ev) => {
             // Check that the user didn't already leave the community before this event arrived
             if !state
                 .data
@@ -99,7 +99,7 @@ fn process_event(event: Event, state: &mut RuntimeState) {
                 state.data.award_achievement_and_notify(Achievement::JoinedCommunity, now);
             }
         }
-        Event::DiamondMembershipPaymentReceived(ev) => {
+        LocalUserIndexEvent::DiamondMembershipPaymentReceived(ev) => {
             let mut awarded = state.data.award_achievement(Achievement::UpgradedToDiamond, now);
 
             if matches!(ev.duration, DiamondMembershipPlanDuration::Lifetime) {
@@ -130,7 +130,7 @@ fn process_event(event: Event, state: &mut RuntimeState) {
                 state.push_user_canister_event(referred_by.into(), UserCanisterEvent::SetReferralStatus(Box::new(status)))
             }
         }
-        Event::NotifyUniquePersonProof(proof) => {
+        LocalUserIndexEvent::NotifyUniquePersonProof(proof) => {
             if state.data.award_achievement(Achievement::ProvedUniquePersonhood, now) {
                 state.data.notify_user_index_of_chit(now);
             }
@@ -143,7 +143,7 @@ fn process_event(event: Event, state: &mut RuntimeState) {
                 )
             }
         }
-        Event::ExternalAchievementAwarded(ev) => {
+        LocalUserIndexEvent::ExternalAchievementAwarded(ev) => {
             state.data.award_external_achievement(ev.name, ev.chit_reward, now);
         }
     }